requirements.txt

fastapi>=0.95.0
uvicorn[standard]>=0.22.0
selenium>=4.15.0
beautifulsoup4>=4.12.0
requests>=2.25.0
reportlab>=4.0.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
PyPDF2>=3.0.0
